#30 RegionRepository의 제네릭 타입 변경

+++ src/main/java/io/wisoft/aoiandregionmatcherv1/repository/RegionRepository.java
@@ -4,8 +4,6 @@ import io.wisoft.aoiandregionmatcherv1.entity.Region;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
-public interface RegionRepository extends JpaRepository<Region, BigInteger> {
+public interface RegionRepository extends JpaRepository<Region, Integer> {
 }
